type-inf-check: add return for 'return' expressions. This notably doesn't include the type for return, which should be 'never'
This is because the only way to currently receive the 'never' type is due to something like:

```lab
def main := fn() {
	let x = return "hello";
};
```

In this example, 'x' will never actually receive a value, so it receives the type 'never'. While not useful in this case, it would be useful in the following:

```lamb
def if_example := fn(xs) -> {
    x := if len(xs) == 0 {
      return 0;
    } else {
      0
    };

    x[0]
};
```

This is of course a contrived example, but shows some use of having 'return' being compatible with any type.