fix(name-resolution): fix forward declaration in during name-resolution
The implementation before this fix declared a new variable for each of the items within a module. This misses that we prematurely create a `Var` for each of the exported items. This would result in items being imported being unaddressable. This can be visualized here:

```
-- File b.lb

export { x }
--       ^ Becomes `Var(x)`

def x: int = 2;
--  ^ Has the value `Var(UNKNOWN)`

-- File a.lb
from "a.lb" import { x };
--                   ^ `Var(x)`

def main
  : fn() -> int { x };
--                ^ Var(x)

```

This fix makes sure that, before we create a `Var` for an item within a module, we first check to see if we can find it in the list of exported items. This was found due to the above imploding during type-checking. Without `Var(x)` not being found in the global scope, there is a missing entry in the global variable map.