fix(type-checking): don't instantiate the type definition of an item
By instantiating the scheme of an item, a bug creeps into the Type-System. All the sudden, the following type-checks:

```
def id
  : fn[x](x) -> x
  = fn(x) -> "sneaky";
```

This type-checks because the type of the function `fn[a](a) -> string` is allowed to be unified with `fn[x](x) -> x`. This produces the type `fn(string) -> string`, but this doesn't produce a type error because references to `id` will use a new instantiation of the scheme. In order to solve this, we need to use the actual type from `scheme` due to the `RigidVar` that it contains. Additionally we need to ensure that when instantiating the type-variables that belong to the type-scheme, and not to a "parent" scheme.

`x` has the scheme `TypeScheme { unbound: {}, ty: Type::Rigid(RigidVar) }` because it is the argument of the parent. Using a reference to `x` would trigger `instantiate`- The issue remains that we can't instantiate `x`, nor would we. So, if the `RigidVar` is one of the unbound variables for that type-scheme, we don't consider it an error. Whether this is the correct choice... we'll see.