better-documentation: add documentation to the top-level items that make up the parsing of a module
While it would also be possible to add doc tests to the finer-grain items of the parser, I don't see a necessity to. The behavior of the rest of the items are covered in the various tests, and those items aren't public facing anyhow. It may be a consideration in the future. This adds enough documentation to give a cursory overview of the parser and it's usage without going overboard. A curious reader can see the tests for more concrete examples of the expression parser.