type-checking: treat function types as whole types
The previous type of type-checking works well for unifying unary functions, however it doesn't handle functions with more than one parameter.

The new type-checking treats the function type as a whole, hence the changes in the unification process. When types are handled with spans, the errors should point to the parameters which have the improper type, or the return expression with the improper type.

fixes #16