add trait for Unifier to easier refactoring/testing
I would like to have a InferenceContext struct which contains the necessary bits required for the type inference, as this should be shared throughout the entire type-inference stages. The issue herein, is that I don't want to have to create a giant struct everytime just so that I can test that unification does what I expect. To work around this, we can introduce traits.