main: fix examples / readme
At the time I put down the project I hadn't yet built in that 'main' should automatically be called, and so none of the main functions were being run.
Due to the grammar of the langauge, only definitions are allowed at the top level, hence the requirement for a 'result' variable. When I inevitably pick
up this project again, that will be one of the first items I should address.